requirements.txt
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
colorama==0.4.6
distlib==0.3.8
factory-boy==3.3.0
Faker==25.9.1
filelock==3.15.4
greenlet==3.0.3
hypothesis==6.103.5
idna==3.7
iniconfig==2.0.0
mimesis==17.0.0
packaging==24.1
platformdirs==4.2.2
playwright==1.44.0
pluggy==1.5.0
pyee==11.1.0
pytest==8.2.2
pytest-base-url==2.1.0
pytest-playwright==0.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
six==1.16.0
sortedcontainers==2.4.0
text-unidecode==1.3
typing_extensions==4.12.2
urllib3==2.2.2
virtualenv==20.26.3
webdriver-manager==4.0.1
